# handover_roadmap/wp-control-center-starter/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: WP Control Center API
  version: 1.0.0
  description: >-
    REST API specification for the WP Control Center admin dashboard, worker
    orchestration layer, and WordPress agent integration.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:3001
    description: Local
security:
  - bearerAuth: []
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List users
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    patch:
      tags: [Users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      tags: [Users]
      summary: Deactivate user
      responses:
        '204':
          description: User deactivated
  /api/sites:
    get:
      tags: [Sites]
      summary: List sites
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/SiteStatus'
        - in: query
          name: connectionStatus
          schema:
            $ref: '#/components/schemas/ConnectionStatus'
        - in: query
          name: hasUpdates
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Sites list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Site'
    post:
      tags: [Sites]
      summary: Create site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '201':
          description: Site created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
  /api/sites/{id}:
    parameters:
      - $ref: '#/components/parameters/SiteId'
    get:
      tags: [Sites]
      summary: Get site detail
      responses:
        '200':
          description: Site detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
    patch:
      tags: [Sites]
      summary: Update site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteRequest'
      responses:
        '200':
          description: Site updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
    delete:
      tags: [Sites]
      summary: Archive site
      responses:
        '204':
          description: Site archived
  /api/sites/{id}/generate-connection-token:
    post:
      tags: [Sites]
      summary: Generate one-time connection token for WordPress agent
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Connection token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
  /api/sites/{id}/rotate-secret:
    post:
      tags: [Sites]
      summary: Rotate site secret keys
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Secret rotated
          content:
            application/json:
              schema:
                type: object
                properties:
                  publicKey:
                    type: string
                  rotatedAt:
                    type: string
                    format: date-time
  /api/sites/{id}/resync:
    post:
      tags: [Sites]
      summary: Queue full site resynchronization job
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '202':
          description: Resync job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobQueueResponse'
  /api/sites/{id}/overview:
    get:
      tags: [Site Inventory]
      summary: Get overview panel data for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Site overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteOverviewResponse'
  /api/sites/{id}/plugins:
    get:
      tags: [Site Inventory]
      summary: List plugins for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Plugin list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plugin'
  /api/sites/{id}/themes:
    get:
      tags: [Site Inventory]
      summary: List themes for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Theme list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Theme'
  /api/sites/{id}/core:
    get:
      tags: [Site Inventory]
      summary: Get WordPress core version status for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Core version info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CoreVersion'
  /api/sites/{id}/uptime:
    get:
      tags: [Monitoring]
      summary: Get recent uptime checks for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - in: query
          name: days
          schema:
            type: integer
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: Uptime checks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UptimeCheck'
  /api/sites/{id}/incidents:
    get:
      tags: [Monitoring]
      summary: Get incidents for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Incident list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
  /api/sites/{id}/analytics:
    get:
      tags: [Analytics]
      summary: Get combined analytics overview for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
        - in: query
          name: startDate
          required: true
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Analytics overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteAnalyticsResponse'
  /api/sites/{id}/audit-logs:
    get:
      tags: [Sites]
      summary: Get audit logs for a site
      parameters:
        - $ref: '#/components/parameters/SiteId'
      responses:
        '200':
          description: Audit log list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLog'
...
  /api/jobs:
    get:
      tags: [Jobs]
      summary: List jobs
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/JobStatus'
        - in: query
          name: siteId
          schema:
            type: string
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
  /api/jobs/{id}:
    parameters:
      - $ref: '#/components/parameters/JobId'
    get:
      tags: [Jobs]
      summary: Get job detail
      responses:
        '200':
          description: Job detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Job'
                  - type: object
                    properties:
                      logs:
                        type: array
                        items:
                          $ref: '#/components/schemas/JobLog'
  /api/jobs/{id}/retry:
    post:
      tags: [Jobs]
      summary: Retry failed job
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Retry queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobQueueResponse'
  /api/jobs/{id}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel queued or running job if supported
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '202':
          description: Cancel requested
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
    SiteId:
      name: id
      in: path
      required: true
      schema:
        type: string
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    UserRole:
      type: string
      enum: [SUPER_ADMIN, ADMIN, OPERATOR, VIEWER]
    EnvironmentType:
      type: string
      enum: [PRODUCTION, STAGING, DEVELOPMENT]
    SiteStatus:
      type: string
      enum: [ACTIVE, DISABLED, ARCHIVED]
    ConnectionStatus:
      type: string
      enum: [PENDING, CONNECTED, DISCONNECTED, ERROR]
    JobStatus:
      type: string
      enum: [QUEUED, RUNNING, SUCCESS, FAILED, CANCELED, TIMED_OUT]
    JobType:
      type: string
      enum:
        - REGISTER_SITE
        - RESYNC_SITE
        - SYNC_SYSTEM_INFO
        - SYNC_PLUGINS
        - SYNC_THEMES
        - SYNC_CORE
        - UPDATE_CORE
        - UPDATE_PLUGIN
        - UPDATE_THEME
        - ACTIVATE_PLUGIN
        - DEACTIVATE_PLUGIN
        - INSTALL_PLUGIN
        - DELETE_PLUGIN
        - DELETE_THEME
        - TOGGLE_MAINTENANCE
        - CLEAR_CACHE
        - OPTIMIZE_DATABASE
        - UPDATE_ROBOTS_TXT
        - UPDATE_HTACCESS
        - UPDATE_PHP_CONFIG
    IncidentStatus:
      type: string
      enum: [OPEN, RESOLVED, MUTED]
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        isActive:
          type: boolean
    Site:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        siteUrl:
          type: string
          format: uri
        environment:
          $ref: '#/components/schemas/EnvironmentType'
        wpVersion:
          type: string
          nullable: true
        phpVersion:
          type: string
          nullable: true
        wpAgentVersion:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/SiteStatus'
        connectionStatus:
          $ref: '#/components/schemas/ConnectionStatus'
        maintenanceMode:
          type: boolean
        ga4PropertyId:
          type: string
          nullable: true
        gscSiteUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSiteRequest:
      type: object
      required: [name, domain, siteUrl]
      properties:
        name:
          type: string
        domain:
          type: string
        siteUrl:
          type: string
          format: uri
        environment:
          $ref: '#/components/schemas/EnvironmentType'
        timezone:
          type: string
        notes:
          type: string
    UpdateSiteRequest:
      type: object
      properties:
        name:
          type: string
        siteUrl:
          type: string
          format: uri
        environment:
          $ref: '#/components/schemas/EnvironmentType'
        status:
          $ref: '#/components/schemas/SiteStatus'
        timezone:
          type: string
        notes:
          type: string
    Plugin:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        slug:
          type: string
        name:
          type: string
        versionInstalled:
          type: string
        versionLatest:
          type: string
          nullable: true
        isActive:
          type: boolean
        isPremium:
          type: boolean
        autoUpdateEnabled:
          type: boolean
        updateAvailable:
          type: boolean
        sourceType:
          type: string
          nullable: true
        installedAt:
          type: string
          format: date-time
          nullable: true
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
    Theme:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        slug:
          type: string
        name:
          type: string
        versionInstalled:
          type: string
        versionLatest:
          type: string
          nullable: true
        isActive:
          type: boolean
        updateAvailable:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
    CoreVersion:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        versionInstalled:
          type: string
        versionLatest:
          type: string
          nullable: true
        updateAvailable:
          type: boolean
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
    Job:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        jobType:
          $ref: '#/components/schemas/JobType'
        targetType:
          type: string
          enum: [SITE, CORE, PLUGIN, THEME, CONFIG, CACHE, DATABASE]
        targetSlug:
          type: string
          nullable: true
        payloadJson:
          type: object
          additionalProperties: true
          nullable: true
        status:
          $ref: '#/components/schemas/JobStatus'
        initiatedByUserId:
          type: string
          nullable: true
        queuedAt:
          type: string
          format: date-time
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
        retryCount:
          type: integer
        errorMessage:
          type: string
          nullable: true
        resultJson:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobLog:
      type: object
      properties:
        id:
          type: string
        jobId:
          type: string
        level:
          type: string
          enum: [INFO, WARN, ERROR, DEBUG]
        message:
          type: string
        contextJson:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
    JobQueueResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          example: queued
    UptimeCheck:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        checkedAt:
          type: string
          format: date-time
        statusCode:
          type: integer
          nullable: true
        responseTimeMs:
          type: integer
          nullable: true
        isUp:
          type: boolean
        errorMessage:
          type: string
          nullable: true
    Incident:
      type: object
      properties:
        id:
          type: string
        siteId:
          type: string
        incidentType:
          type: string
          enum: [UPTIME, RESPONSE_TIME, CONNECTION]
        severity:
          type: string
          enum: [INFO, WARNING, CRITICAL]
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/IncidentStatus'
        summary:
          type: string
          nullable: true
        metadataJson:
          type: object
          additionalProperties: true
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AnalyticsPoint:
      type: object
      properties:
        metricDate:
          type: string
          format: date
        sessions:
          type: integer
          nullable: true
        users:
          type: integer
          nullable: true
        pageviews:
          type: integer
          nullable: true
        impressions:
          type: integer
          nullable: true
        clicks:
          type: integer
          nullable: true
        ctr:
          type: number
          format: float
          nullable: true
        avgPosition:
          type: number
          format: float
          nullable: true
    TopPage:
      type: object
      properties:
        pagePath:
          type: string
        pageTitle:
          type: string
          nullable: true
        sessions:
          type: integer
          nullable: true
        pageviews:
          type: integer
          nullable: true
        impressions:
          type: integer
          nullable: true
        clicks:
          type: integer
          nullable: true
        ctr:
          type: number
          format: float
          nullable: true
        avgPosition:
          type: number
          format: float
          nullable: true
    AnalyticsSeriesResponse:
      type: object
      properties:
        source:
          type: string
          enum: [GA4, GSC]
        series:
          type: array
          items:
            $ref: '#/components/schemas/AnalyticsPoint'
        topPages:
          type: array
          items:
            $ref: '#/components/schemas/TopPage'
    SiteAnalyticsResponse:
      type: object
      properties:
        ga4:
          $ref: '#/components/schemas/AnalyticsSeriesResponse'
        gsc:
          $ref: '#/components/schemas/AnalyticsSeriesResponse'
    SiteOverviewResponse:
      type: object
      properties:
        site:
          $ref: '#/components/schemas/Site'
        core:
          $ref: '#/components/schemas/CoreVersion'
        pluginSummary:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            updatesAvailable:
              type: integer
        themeSummary:
          type: object
          properties:
            total:
              type: integer
            updatesAvailable:
              type: integer
        monitoring:
          type: object
          properties:
            isUp:
              type: boolean
            averageResponseTimeMs:
              type: integer
            openIncidents:
              type: integer
        recentJobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
    AuditLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
          nullable: true
        siteId:
          type: string
          nullable: true
        action:
          type: string
        entityType:
          type: string
          nullable: true
        entityId:
          type: string
          nullable: true
        payloadJson:
          type: object
          additionalProperties: true
          nullable: true
        result:
          type: string
          enum: [SUCCESS, FAILURE, PARTIAL]
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...